+++ sdk/core/azure-core/test/ut/datetime.cpp
@@ -168,7 +168,7 @@ TEST(DateTime, EmittingTimeCorrectDay)
 }
 
 namespace {
-void TestRfc1123IsTimeT(char const* str, long long t)
+void TestRfc1123IsTimeT(char const* str, int64_t t)
 {
   auto const dt = DateTime::Parse(str, DateTime::DateFormat::Rfc1123);
   auto interval = (((dt - year2001).GetNanoseconds()) / 100)
@@ -196,7 +196,7 @@ TEST(DateTime, ParseTimeRfc1123AcceptsEachDay)
 TEST(DateTime, ParseTimeRfc1123BoundaryCases)
 {
   TestRfc1123IsTimeT("01 Jan 1970 00:00:00 GMT", 0);
-  TestRfc1123IsTimeT("19 Jan 2038 03:14:06 GMT", std::numeric_limits<int>::max() - 1);
+  TestRfc1123IsTimeT("19 Jan 2038 03:14:06 GMT", static_cast<int64_t>(std::numeric_limits<int>::max()) - 1);
   TestRfc1123IsTimeT("19 Jan 2038 03:13:07 -0001", std::numeric_limits<int>::max());
   TestRfc1123IsTimeT("19 Jan 2038 03:14:07 -0000", std::numeric_limits<int>::max());
   TestRfc1123IsTimeT("14 Jan 2019 23:16:21 +0000", 1547507781);
